test(parametric): isolate otlp trace export tests
