+++ manifests/golang.yml
@@ -1550,7 +1550,7 @@ manifest:
   : "irrelevant (\"Go tracer decided to always set _dd1.sr.eausr: 1 for truthy analytics.event inputs, else 0\")"
   tests/parametric/test_otel_span_with_baggage.py::Test_Otel_Span_With_Baggage: missing_feature
   tests/parametric/test_otel_tracestate_sampling.py::Test_OtelTracestateSampling: v2.11.0-dev
-  tests/parametric/test_otel_tracestate_sampling.py::Test_OtlpTracestateSampling: v2.11.0-dev
+  tests/parametric/test_otlp_trace_export.py::Test_OtlpTracestateSampling: v2.11.0-dev
   tests/parametric/test_otlp_trace_metrics.py: v2.11.0-dev
   tests/parametric/test_otlp_trace_metrics.py::Test_FR06_Otel_Resource_Attributes::test_fr06_15_default_service_on_data_point: missing_feature (default service not on data point)
   tests/parametric/test_otlp_trace_metrics.py::Test_FR06_Otel_Span_Attributes::test_fr06_2_span_kind: missing_feature (span.kind is emitted as the raw Datadog value, e.g. "server", not the SMC's SPAN_KIND_* convention)
