+++ tests/parametric/test_otel_tracestate_sampling.py
@@ -1,12 +1,6 @@
 """Cross-tracer conformance for OpenTelemetry consistent probability sampling (ot.th / ot.rv) on the wire."""
 
-import base64
-import json
-import time
-
 import pytest
-from google.protobuf.json_format import MessageToDict
-from opentelemetry.proto.collector.trace.v1.trace_service_pb2 import ExportTraceServiceRequest
 
 from utils import features, scenarios
 from utils.dd_constants import SamplingPriority
@@ -214,20 +208,6 @@ def _library_env(rate: float) -> dict[str, str]:
     }
 
 
-def _otlp_library_env() -> dict[str, str]:
-    return {
-        **_library_env(0.1),
-        "OTEL_EXPORTER_OTLP_TRACES_PROTOCOL": "http/json",
-        "OTEL_TRACES_EXPORTER": "otlp",
-    }
-
-
-@pytest.fixture
-def otlp_library_env(library_env: dict[str, str], test_agent: TestAgentAPI) -> dict[str, str]:
-    library_env["OTEL_EXPORTER_OTLP_TRACES_ENDPOINT"] = f"http://{test_agent.container_name}:4318/v1/traces"
-    return library_env
-
-
 def _make_child_headers(test_library: APMLibrary, headers: list[tuple[str, str]]) -> dict[str, str]:
     return test_library.dd_make_child_span_and_get_headers(headers)
 
@@ -242,37 +222,6 @@ def _make_root_manual_keep_headers(test_library: APMLibrary) -> dict[str, str]:
     with test_library.dd_start_span("name") as span:
         span.manual_keep()
         return {key.lower(): value for key, value in test_library.dd_inject_headers(span.span_id)}
-
-
-def _otlp_trace_requests(test_agent: TestAgentAPI) -> list[dict]:
-    return [request for request in test_agent.otlp_requests() if request["url"].endswith("/v1/traces")]
-
-
-def _wait_for_otlp_trace_requests(test_agent: TestAgentAPI) -> list[dict]:
-    for _ in range(80):
-        requests = _otlp_trace_requests(test_agent)
-        if requests:
-            return requests
-        time.sleep(0.1)
-    raise AssertionError("No OTLP trace export request received")
-
-
-def _decode_otlp_trace_request(request: dict) -> dict:
-    body = base64.b64decode(request["body"])
-    headers = {key.lower(): value for key, value in request["headers"].items()}
-    if "json" in headers.get("content-type", ""):
-        return json.loads(body.decode("utf-8"))
-    return MessageToDict(ExportTraceServiceRequest.FromString(body), preserving_proto_field_name=False)
-
-
-def _otlp_spans(requests: list[dict]) -> list[dict]:
-    spans: list[dict] = []
-    for request in requests:
-        body = _decode_otlp_trace_request(request)
-        for resource_span in body.get("resourceSpans", []):
-            for scope_span in resource_span.get("scopeSpans", []):
-                spans.extend(scope_span.get("spans", []))
-    return spans
 
 
 SAMPLE_RATE_VECTORS = [
@@ -557,62 +506,3 @@ class Test_OtelTracestateSampling:
         traceparent, tracestate = get_tracecontext(headers)
         assert _parse_ot(tracestate) == {"rv": expected_rv, "th": expected_th}
         assert (traceparent.trace_flags == "01") is expected_sampled
-
-
-@features.otel_api
-@scenarios.parametric
-class Test_OtlpTracestateSampling:
-    @pytest.mark.parametrize("library_env", [_otlp_library_env()])
-    def test_otlp_carries_probability_sampling_ot(
-        self,
-        otlp_library_env: dict[str, str],  # noqa: ARG002
-        test_agent: TestAgentAPI,
-        test_library: APMLibrary,
-    ) -> None:
-        """B1: a probability sampling decision is carried on the exported OTLP span's tracestate as ot=rv:...;th:...
-
-        See APMAPI-2172.
-        """
-        with test_library as library:
-            with library.dd_extract_headers_and_make_child_span(
-                "name",
-                [
-                    ("x-datadog-trace-id", str(FORWARD_TRACE_ID)),
-                    ("x-datadog-parent-id", str(FORWARD_TRACE_ID)),
-                ],
-            ):
-                pass
-            assert library.dd_flush()
-
-        spans = _otlp_spans(_wait_for_otlp_trace_requests(test_agent))
-        assert spans, "No span found in the OTLP trace export"
-        ot = _parse_ot(spans[0].get("traceState", ""))
-        assert ot.get("rv") == FORWARD_RV
-        assert ot.get("th") == FORWARD_TH
-
-    @pytest.mark.parametrize("library_env", [_otlp_library_env()])
-    def test_otlp_forwards_inherited_ot(
-        self,
-        otlp_library_env: dict[str, str],  # noqa: ARG002
-        test_agent: TestAgentAPI,
-        test_library: APMLibrary,
-    ) -> None:
-        """B2: an inherited ot=rv:...;th:... is forwarded unchanged onto the exported OTLP span's tracestate.
-
-        See APMAPI-2172.
-        """
-        with test_library as library:
-            with library.dd_extract_headers_and_make_child_span(
-                "name",
-                [
-                    ("traceparent", _traceparent(FORWARD_TRACE_ID, sampled=True)),
-                    ("tracestate", f"dd=s:2;t.dm:-3,ot=rv:{FORWARD_RV};th:{FORWARD_TH}"),
-                ],
-            ):
-                pass
-            assert library.dd_flush()
-
-        spans = _otlp_spans(_wait_for_otlp_trace_requests(test_agent))
-        assert spans, "No span found in the OTLP trace export"
-        ot = _parse_ot(spans[0].get("traceState", ""))
-        assert ot == {"rv": FORWARD_RV, "th": FORWARD_TH}
